+++ packages/terra-status-view/CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ----------
 
+### Fixed
+* Status View glyph does not always displays when it should
+
 3.3.0 - (January 26, 2019)
 ------------------
 ### Changed